feat: null 값 및 빈 값 입력 시 예외 처리 기능 구현

+++ src/main/java/roomescape/domain/reservations/application/ReservationService.java
@@ -5,20 +5,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.reservations.entity.Reservation;
 import roomescape.domain.reservations.entity.ReservationTime;
-import roomescape.domain.reservations.infrastructure.ReservationJdbcTemplateRepository;
-import roomescape.domain.reservations.infrastructure.ReservationTimeJdbcTemplateRepository;
+import roomescape.domain.reservations.infrastructure.ReservationRepository;
+import roomescape.domain.reservations.infrastructure.ReservationTimeRepository;
 import roomescape.domain.reservations.presentation.dto.ReservationRequest;
 import roomescape.domain.reservations.presentation.dto.ReservationResponse;
 
 @Service
 public class ReservationService {
 
-    private final ReservationJdbcTemplateRepository reservationRepository;
-    private final ReservationTimeJdbcTemplateRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
 
     public ReservationService(
-            ReservationJdbcTemplateRepository reservationRepository,
-            ReservationTimeJdbcTemplateRepository reservationTimeRepository
+            ReservationRepository reservationRepository,
+            ReservationTimeRepository reservationTimeRepository
     ) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
@@ -26,6 +26,8 @@ public class ReservationService {
 
     @Transactional
     public ReservationResponse saveReservation(ReservationRequest request) {
+        validateSaveRequest(request);
+
         ReservationTime time = reservationTimeRepository.findById(request.timeId())
                 .orElseThrow(IllegalArgumentException::new);
         Reservation reservation = Reservation.of(
@@ -43,6 +45,24 @@ public class ReservationService {
     }
 
     public void deleteReservation(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException();
+        }
         reservationRepository.deleteById(id);
+    }
+
+    private void validateSaveRequest(ReservationRequest request) {
+        if (request == null) {
+            throw new IllegalArgumentException();
+        }
+        if (request.name() == null || request.name().trim().isBlank()) {
+            throw new IllegalArgumentException();
+        }
+        if (request.date() == null) {
+            throw new IllegalArgumentException();
+        }
+        if (request.timeId() == null) {
+            throw new IllegalArgumentException();
+        }
     }
 }
