+++ src/main/java/roomescape/domain/reservations/application/ReservationTimeService.java
@@ -4,21 +4,23 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.reservations.entity.ReservationTime;
-import roomescape.domain.reservations.infrastructure.ReservationTimeJdbcTemplateRepository;
+import roomescape.domain.reservations.infrastructure.ReservationTimeRepository;
 import roomescape.domain.reservations.presentation.dto.ReservationTimeRequest;
 import roomescape.domain.reservations.presentation.dto.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
 
-    private final ReservationTimeJdbcTemplateRepository reservationTimeRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeJdbcTemplateRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
     @Transactional
     public ReservationTimeResponse saveTime(ReservationTimeRequest request) {
+        validateSaveRequest(request);
+
         ReservationTime reservationTime = ReservationTime.of(
                 null,
                 request.startAt()
@@ -32,6 +34,18 @@ public class ReservationTimeService {
     }
 
     public void deleteTime(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException();
+        }
         reservationTimeRepository.deleteById(id);
+    }
+
+    private void validateSaveRequest(ReservationTimeRequest request) {
+        if (request == null) {
+            throw new IllegalArgumentException();
+        }
+        if (request.startAt() == null) {
+            throw new IllegalArgumentException();
+        }
     }
 }
